[jancy] TypedefShadowType
when generating documentation, we want to keep typedefs in tact instead of substituting the actual type. at the same time, during compilation, we HAVE TO substitute -- otherwise it's unnecessary inefficient and besides, too much has to be re-written. as a compromise solution, i decided to introduce a special type TypedefShadowType which keeps symbolic information about typedef and is used only when generating documentation.