[jnc_io_base] added io.Socket.m_isSocketEventEnabled property instead of io.Socket.firePendingEvents (which was never really implemented)
the problem being solved is when we 'accept' a connection some events may fire BEFORE we have a chance to assign an event handler, so we either lose them or have to write an additional block for checking and handling these events after 'accept'.

the better approach seems to be to always 'accept' with DISABLED events; then we set our infrastructure up and assign an event handler, and when we are ready, we ENABLE events -- all the pending events will be replayed in the sequence they were originally fired and we will handle them normally.