[jnc_rt] allow wait regions within no-collect regions (just trace a warning)
that's not really an indication of a bug or even a problem. just something to be careful about (if we enter no-collect region and wait for a long time, collections are completely disabled until we finish waiting).

but in most cases it introduces no problems whatsoever (e.g. sys.sleep (1000))