[jnc_doc] added recommendation to add -fvisibility=hidden when building LLVM
otherwise, linker may add a subset of LLVM symbols to '.dynsym' table, which in
turn leads to crashes of IO Ninja on Linux systems with mesa OpenGL
implementation.

the issue is really hard to track, cause the actual subset of added symbols is
build-machine-dependant and it MAY or MAY NOT lead to a crash on a particular
Linux.