[all] re-organize source folders
i'm tired of waiting for the support for source_groups in either QT creator,
NetBeans or Clion -- guess it's never gonna happen. the best i can do is move
files around and use dirs instead of source_groups.

that's not nearly as flexible as virtual source_groups and incurs creating
additional include dirs with re-directing headers, but oh well..

without folders (with a planar representation of jnc_ct) it's a major pain to
modify Jancy compiler on Linux.