[jnc_ct][jnc_ext] remove JNC_LIB_FORCED_ macros -- instead, do it via normal JNC_LIB_SOURCE_FILE and then JNC_LIB_IMPORT
it was a minor and unnecessary optimization which just complicated the API).

also, ignore '.'-prefixed files in dynamic extensions -- this allows to specify which files are to be force-imported by the IDE. currently, this feature is used in static extensions only -- all the dynamic extensions which do need a forced import have a single file, which should be treated as the forced import anyway (why would you NOT want to import the only source file?).