[jnc_ct] major update: redesign of dual modifiers, implemented dual modifier 'cmut' (container mutability).
originally, all dual types carried a reference to the anchor namespace; the duality was resolved at the moment of application (during a const-check or during a event-access-check). this model simply does not work for 'cmut'-style duality; besides, this implementation turned out to be pretty heavy (extra fields and lots of extra code for anchor namespace handling all around).

the better approach to duality seems to be folding on member operator; that is, the dual type remains dual and carries no extra anchor namespace fields and whatnot -- instead, the anchor namespace vs current namespace relation is checked and the dual type is folded into a non-dual type inside member operator; container mutability for 'cmut' can also be checked with this approach (member operator accepts a pointer to the container).

currently, the implementation does not cover 100% of use-cases (e.g. dual pointer return values or dual properties); but it works for fields, i.e. the most important use-case of dual modifiers.