[jnc_ct][jnc_rt][jnc_rtl][jnc_std] re-engineered dynamic layout lookup implementation
originally, each box had a pointer to a dynamic layout. firstly, that's too wasteful -- dynamic layouts are not normally used too often, but space is taken from each and every box.  secondly, it results in alignment issues on 32-bits, so extra padding is needed -- which makes it even more wasteful.

i decided to remove the direct pointer and replace it with a hash table map maintained by the gc heap.