[jnc_ct] major update/bugfix: StdFunc_AllocateXxx/StdFunc_TryAllocateXxx should return StdType_AbstractDataPtr, not StdType_DataPtrStruct
StdType_DataPtrStruct is non-gc-root, therfore, everything was relying on manually marking the result as gc root -- which was not always the case. E.g., lifting out-of-place large variant values; passing a few newly allocated arguments: foo (new Struct, new Struct); etc.

it's much more reliable to mark the result of data-allocating funcs, just like we do with class-allocating funcs.

for this to work properly, we need to globally signal we are lifting a variable so that GcShadowStackMgr::markGcRoot (and consequently GcShadowStackMgr::openFrameMap) know the target scope -- we do that with VariableMgr::getCurrentLiftedStackVariable

this change renders extra markGcRoot in VariableMgr::allocateHeapVariable redundant; also, there's no need in extra Scope* argument in GcShadowStackMgr::markGcRoot (we use VariableMgr::getCurrentLiftedStackVariable)