[jnc_rt] fix: invokePrevSignalHandler did not really work
raise does not necessarily call signal handler immediately, so we need
to do that ourselves (or restore, raise, and leave -- if we can't)

the whole thing with invoking prev signal handler is an attempt to
ensure jancy library will not break other 3rd party libraries which may
also install signal handlers. that's not really an issue in ioninja.