[jnc_io] critical bugfix: lock inside markOpaqueGcRoots
io threads are NOT permanently added as GC mutator threads -- they are added on as-needed basis, whenever we are ready to fire a notification. as such, opaque pointer lists may get mutated while GC makes a mark/sweep.

i found this bug while investigating a rare opportunistic crash in io::SocketAddressResolver -- turned out, GC squeezed its mark stage in-between removing a request from pending list and adding it into active list; the closure object of completion routine was not marked and was sweeped, leaving a request with a dangling closure pointer.

another important thing is to check whether markOpaqueGcRoots is called BEFORE constructor -- it very well might happen (when we allocate a class, we prime it and then add its box to the list, after which moment it will be traced)