[jnc_io] accepted sockets/named-pipes and usb IN-endpoints should be initially suspendable
this way we can configure those prior to start reading (i.e. settings such as m_readBlockSize, m_readParallelism, etc). in case of sockets and files we could configure accepted socket via the parent socket (prior to accept), but io.UsbEndpoint just doesn't (and shouldn't) have those settings. the most natural way would be to create socket/file/endpoint suspended, configure it and then "unsuspend".

in order to maintain backward compatibility (and to make it work "out-of-the-box" in most cases), we add "isSuspended = false" argument to accept() and openEndpoint() methods