[jnc_io_pcap] use signals and pcap_breakloop on close
as it turns out, some platforms don't support the timeout for live
capture in the same way as WinPcap does; so we need a way to interrupt
pcap_next_ex on close -- in case no packet ever shows up on the adapter.

since pcap_next_ex calls pcap_loop/pcap_dispatch internally, we can use
pcap_breakloop; the last quirk is that it has to be called in the same
thread as pcap_next_ex. hence, signals and pthread_kill.