[jnc_api][jnc_ct][jnc_edit] added support for auto-completing full function declarations
this is really useful for generating templates for some predefined entry-points. for instance, io ninja script now support log listeners via the 'onLogRecord' entry point which takes 4 parameters.
now we can mark a forward-decl for 'onLogRecord' with [ autoCompleteDecl ], and when a user presses ctrl+space and selects onLogRecord, a valid template with correct parameters is generated.