[jnc_api][jnc_rt] added jnc_setJmp/jnc_longJmp (ensure no unwinding during SJLJ)
SJLJ-unwinding from LLVM JITted code causes problems on certain combinations of CRT/CPU/optimizations

Jancy will use a simple SJLJ as to avoid those issues. currently, it's Windows-only, inline ASM on x86,
and a bit of a hack on AMD64. in the futute it makes sense to properly compile SJLJ code using NASM on all platforms