[jnc_ct] refactor: JIT takes ownership of the module, so we better delete JIT and all LLVM stuff together
here's where it caused problems: in ioninja packet templates, we tried to get to State_Parsed first, then we calculated attributes (which caused JIT to be created to get static variable addresses), and then we tried to get to State_Compiled when we need to run actions. Compile could fail, causing clearLlvm -- but after that delete m_jit caused a crash.