[jnc_rt] fix: m_stats.m_currentAllocSize was calculated incorrectly
originally, we were bumping m_stats.m_currentAllocSize on allocations, and subtracted the freed size on collect.
freed size was calculated based on box types and element count, but formatting literals adjust its validator range-end
to exclude unused buffer tail -- this adjustment was not accounted for.
the new alg is more straightforward -- we count the marked box size, not freed size (which we calculate in the end)